+++ db/migrate/20230608185139_create_works.rb
@@ -3,7 +3,7 @@ class CreateWorks < ActiveRecord::Migration[7.0]
     create_table :works, id: false, primary_key: :id do |t|
       t.integer :id, null: false
       t.string :nome
-      t.integer :data
+      t.bigint :data
       t.string :descrizione
       t.string :immagini
       t.integer :idstanza
